backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
yfinance>=0.2.40
aiosqlite>=0.20.0
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
slowapi>=0.1.9
